+++ src/main/java/com/example/tuFondaOnline/service/OrdenService.java
@@ -3,7 +3,9 @@ package com.example.tuFondaOnline.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.example.tuFondaOnline.model.Orden;
 import com.example.tuFondaOnline.repository.OrdenRepository;
@@ -13,7 +15,8 @@ import jakarta.transaction.Transactional;
 @Service
 @Transactional
 public class OrdenService {
-     @Autowired
+    
+    @Autowired
     private OrdenRepository ordenRepository;
     
     public List<Orden> findAll(){
@@ -21,7 +24,7 @@ public class OrdenService {
     }
 
     public Orden findById(Long id){
-        return ordenRepository.findById(id).get();
+        return ordenRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Orden no encontrada"));
     }
 
     public Orden save(Orden orden){
@@ -29,6 +32,15 @@ public class OrdenService {
     }
 
     public void delete(Long id){
+        if (!ordenRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se puede eliminar, ID no existe");
+        }
         ordenRepository.deleteById(id);
+    }
+
+   
+    public List<Orden> findByUsuarioEmail(String email) {
+        
+        return ordenRepository.findByUsuario_Email(email);
     }
 }
