+++ src/main/java/com/example/tuFondaOnline/service/ProductoService.java
@@ -3,7 +3,9 @@ package com.example.tuFondaOnline.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.example.tuFondaOnline.model.Producto;
 import com.example.tuFondaOnline.repository.ProductoRepository;
@@ -22,14 +24,19 @@ public class ProductoService {
     }
 
     public Producto findById(Long id){
-        return productoRepository.findById(id).get();
+        return productoRepository.findById(id)
+            
+            .orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Producto no encontrado con ID: " + id));
     }
-
     public Producto save(Producto producto){
         return productoRepository.save(producto);
     }
 
-    public void delete(Long id){
+    public void deleteById(Long id){
+        if (!productoRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se puede eliminar, ID no existe");
+        }
         productoRepository.deleteById(id);
     }
